refactor: update environment variable naming for API integration and enhance session management imports for consistency
